#3204 Improve task creation performance
